requirements.txt
numpy==2.2.4
langchain==0.3.13
sentence-transformers==4.1.0
transformers==4.51.3
langchain-community==0.3.13
faiss-cpu==1.10.0
langchain-chroma==0.2.3
einops==0.8.1
